Fix storing of lan log segfault (#443)
There is a bug report in e-mail (via mailing list), where the reporter explained the situation: if he uses Tlf on two nodes then when he enter a QSO the receiver node exists with SEGFAULT.

The problem is that the function parse_qso() is called two times while accepting the message from the other node, and it uses the buffer variable directly. When the function is called first time, the used strtok_r() chain "eats" the buffer, and at the second call it uses an empty buffer.

Making a copy of the string and passing that to the first call fixes the problem. (See discussion in PR #441)