Hamlib rotator support (#455)
This adds rotator support. Configuration is done exactly like rig
support with ROTATOR_CONTROL, ROTMODEL, ROTSPEED and ROTPORT.

Two new keybindings are introduced: ^ (think "arrow") rotates towards
the country of the current call, and & (the key right of ^ on US
keyboards) rotates to the long-path direction. If the current call is
from the own country, no rotation is done.